Refactor of search result components
Refactored search result components into individual files.
