--- src/components/Business.ts
@@ -1,10 +1,0 @@
-type Business = {
-    id: any;
-    businessName: string;
-    distance: string;
-    futureProofRating: Number;
-    userRating: Number;
-    logoURL: string
-  };
-
-  export default Business
